implement step 1

+++ src/main/scala/model/Fruit.scala
@@ -2,12 +2,12 @@ package model
 
 sealed trait Fruit {
 
-  val price: Double
+  val price: BigDecimal
 
 }
 
 case object Apple extends Fruit {
-  val price = 0.60
+  val price = 0.6
 }
 case object Orange extends Fruit {
   val price = 0.25
